form divided into two separate forms, additional DailyNormaCalculator component added,  yup for calculator  is set up
