+++ src/components/DailyNormaCalculator/DailyNormaCalculator.styled.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Error = styled.p`
+  color: var(--secondary-red);
+  font-size: 14px;
+  line-height: 1.29;
+  position: absolute;
+  bottom: -17px;
+`;
